requirements.txt
Flask==2.1.3
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
PyJWT==2.4.0
python-dotenv==1.0.0
SQLAlchemy==1.4.39
psycopg2-binary==2.9.5
